Splash and home page routes and install json server
