+++ src/components/Rooms/Rooms.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Rooms = () => (
+  <h2>Rooms</h2>
+);
+
+export default Rooms;
