+++ src/components/Splash/Splash.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Splash = () => (
+  <h2>Splash</h2>
+);
+
+export default Splash;
